updated code to the new version post Arduino 1.0
